Take IHttpContextAccessor out of services

+++ CraftIt.Api/Controllers/CommentControllers.cs
@@ -18,12 +18,14 @@ namespace CraftIt.Api.Controllers
 
         private readonly IProductRepository _productRepository;
         private readonly ICommentService _commentService;
+        private readonly IUserService _userService;
     
 
-        public CommentController(IProductRepository productRepository, ICommentService commentService)
+        public CommentController(IProductRepository productRepository, ICommentService commentService,IUserService userService)
         {
             _productRepository = productRepository;
             _commentService = commentService;
+            _userService = userService;
         }
 
 
@@ -31,7 +33,11 @@ namespace CraftIt.Api.Controllers
         [HttpPost]
         public IActionResult Post([FromBody] CommentCreationDto comment)
         {
-            _commentService.AddComment(comment);
+            var user = _userService.GetById(int.Parse(User.Identity.Name));
+
+            if(user == null) return Unauthorized();
+
+            _commentService.AddComment(comment, user);
 
             if(!_commentService.Save()) throw new Exception("Failed to post new comment");
 
@@ -41,6 +47,13 @@ namespace CraftIt.Api.Controllers
         [HttpPut]
         public IActionResult Put([FromBody] CommentUpdateDto comment)
         {
+
+            var commentInDb = _commentService.GetComment(comment.CommentId);
+
+            if(commentInDb == null) return NotFound();
+
+            if(commentInDb.User.Id != int.Parse(User.Identity.Name)) return Unauthorized();
+
             _commentService.UpdateComment(comment);
 
             if(!_commentService.Save()) throw new Exception("Failed to update comment");
@@ -51,7 +64,14 @@ namespace CraftIt.Api.Controllers
         [HttpDelete]
         public IActionResult Delete(int id){
 
+            var commentInDb = _commentService.GetComment(id);
+
+            if(commentInDb == null) return NotFound();
+
+            if(commentInDb.User.Id != int.Parse(User.Identity.Name)) return Unauthorized();
+
             _commentService.DeleteComment(id);
+
             if(!_commentService.Save()) throw new Exception("Failed to delete comment");
 
             return Ok();
